+++ ext/sqlite3/database.h
@@ -9,6 +9,7 @@ struct _sqlite3Ruby {
     int stmt_timeout;
     struct timespec stmt_deadline;
     rb_pid_t owner;
+    int flags;
 };
 
 typedef struct _sqlite3Ruby sqlite3Ruby;